# pmcluster/docs/openapi.yaml
openapi: 3.0.3
info:
  title: pmcluster daemon API
  description: |
    REST + webhook API for the pmcluster Docker Swarm control plane.

    All `/api/*` routes require a Bearer token (an API user; see `POST /api/api_keys`
    to mint one). The webhook receiver `/webhook/{source}` is intentionally NOT
    Bearer-authenticated — its auth is an HMAC-SHA256 signature computed over the
    (timestamp + body) with the source's shared secret. API users are minted with
    `pmcluster user create` or `POST /api/api_keys`; webhook sources with
    `pmcluster webhook add` or `POST /api/webhooks`.
  version: "0.2.0"
servers:
  - url: https://pmcluster.example.com
    description: Deployed daemon behind the edge gateway
  - url: http://127.0.0.1:PORT
    description: Local `pmcluster serve`
paths:
...
  /api/tls/hosts:
    get:
      summary: List per-host TLS certificates
      description: |
        Per-host certificates live in the SAME table and flow as the cluster's
        own certificate: the PEM bytes are versioned Swarm secrets
        (hostcert-<host>_vNNN / hostkey-<host>_vNNN) and this list returns the
        DB metadata for every row except the cluster's own domain (which is
        served by GET /api/tls/site). Nothing is ever read from the manager's
        filesystem.
      operationId: listHostCerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Host certificate list
          content:
            application/json:
              schema:
                type: object
                properties:
                  hosts:
                    type: array
                    items:
                      $ref: "#/components/schemas/HostCert"
  /api/tls/hosts/{host}:
    put:
      summary: Store a per-host cert/key (as PEM text) and refresh Traefik
      description: |
        Cert and key arrive as PEM TEXT in the JSON body — never as a file upload.
        Only reachable when per-host TLS is wired (HostCertService).
      operationId: putHostCert
      security:
        - bearerAuth: []
      parameters:
        - name: host
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cert, key]
              properties:
                cert:
                  type: string
                key:
                  type: string
      responses:
        "200":
          description: Stored (and Traefik refreshed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostCert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          description: Stored but Traefik refresh failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Remove a per-host certificate
      operationId: removeHostCert
      security:
        - bearerAuth: []
      parameters:
        - name: host
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
  /api/webhooks:
    get:
      summary: List webhook sources (without revealing secrets)
      operationId: listWebhooks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Webhook source list
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Webhook"
    post:
      summary: Create a webhoook source (shared secret returned once)
      description: The HMAC shared secret is generated on the daemon, stored encrypted,
        and returned in the response body ONCE. It can never be read again.
      operationId: createWebhook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Created. `secret` is returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Source already exists
  /api/webhooks/{source}:
    delete:
      summary: Revoke a webhook source
      operationId: deleteWebhook
      security:
        - bearerAuth: []
      parameters:
        - name: source
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Revoked (shared secret invalidated)
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: pmc_<token_id>_<secret>
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    HostCert:
      type: object
      properties:
        host:
          type: string
        sans:
          type: array
          items:
            type: string
        not_after:
          type: string
          format: date-time
        not_before:
          type: string
          format: date-time
        cert_secret:
          type: string
        key_secret:
          type: string
        cert_hash:
          type: string
        key_hash:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Webhook:
      type: object
      properties:
        source:
          type: string
        description:
          type: string
        created_at:
          type: integer
          format: int64
        last_used_at:
          type: integer
          format: int64
    WebhookCreated:
      type: object
      properties:
        source:
          type: string
        secret:
          type: string
...